AssaySampleLookupContext
