+++ api/src/org/labkey/api/assay/AbstractAssayProvider.java
@@ -2114,121 +2114,4 @@ public abstract class AbstractAssayProvider implements AssayProvider
             }
         }
     }
-
-    @Override
-    public void syncSampleLookupLineage(
-        Container container,
-        User user,
-        ExpRun run,
-        TableInfo table,
-        SimpleFilter scopeFilter,
-        Collection<FieldKey> sampleLookups,
-        BatchValidationException errors
-    )
-    {
-        if (sampleLookups.isEmpty())
-            return;
-
-        var inputContext = computeMaterialInputs(container, user, table, scopeFilter, sampleLookups, errors);
-
-        if (errors.hasErrors())
-            return;
-
-        var addedMaterialInputs = inputContext.materialInputs;
-        var lineageRoles = inputContext.lineageRoles;
-        var removedMaterialInputs = new HashSet<Integer>();
-
-        for (var materialEntry : run.getMaterialInputs().entrySet())
-        {
-            // This material input is sourced from a different column or other source. Skip it.
-            if (!lineageRoles.contains(materialEntry.getValue()))
-                continue;
-
-            var material = materialEntry.getKey();
-            if (addedMaterialInputs.containsKey(material))
-                addedMaterialInputs.remove(material);
-            else
-                removedMaterialInputs.add(material.getRowId());
-        }
-
-        if (!removedMaterialInputs.isEmpty())
-        {
-            var pa = run.getInputProtocolApplication();
-            if (pa != null)
-                pa.removeMaterialInputs(user, removedMaterialInputs);
-        }
-
-        if (!addedMaterialInputs.isEmpty())
-        {
-            var pa = run.getInputProtocolApplication();
-            if (pa != null)
-            {
-                for (var entry : addedMaterialInputs.entrySet())
-                    pa.addMaterialInput(user, entry.getKey(), entry.getValue());
-            }
-        }
-    }
-
-    private record InputContext(Map<ExpMaterial, String> materialInputs, Set<String> lineageRoles) {}
-
-    private InputContext computeMaterialInputs(
-        Container container,
-        User user,
-        TableInfo table,
-        SimpleFilter scopeFilter,
-        Collection<FieldKey> sampleLookups,
-        BatchValidationException errors
-    )
-    {
-        var domain = table.getDomain();
-        if (domain == null)
-        {
-            errors.addRowError(new ValidationException(String.format("Failed to resolve domain for table \"%s\".", table.getName())));
-            return new InputContext(emptyMap(), emptySet());
-        }
-
-        var materialInputs = new HashMap<ExpMaterial, String>();
-        var lineageRoles = new HashSet<String>();
-        var columns = QueryService.get().getColumns(table, sampleLookups);
-        var columnContext = new HashMap<ColumnInfo, Pair<DomainProperty, ExpSampleType>>();
-        var remapCache = new RemapCache();
-        var materialsCache = new LRUMap<Integer, ExpMaterial>(1_000);
-
-        for (ColumnInfo column : columns.values())
-        {
-            var dp = domain.getPropertyByURI(column.getPropertyURI());
-            var sampleType = ExperimentService.get().getLookupSampleType(dp, container, user);
-            columnContext.put(column, Pair.of(dp, sampleType));
-        }
-
-        new TableSelector(table, columns.values(), scopeFilter, null).forEachResults(results -> {
-            for (var column : columns.values())
-            {
-                var context = columnContext.get(column);
-
-                try
-                {
-                    // TODO: We likely need to elevate permissions or use a service user here
-                    var expMaterial = ExperimentService.get().findExpMaterial(container, user, column.getValue(results), context.second, remapCache, materialsCache);
-
-                    if (expMaterial != null)
-                    {
-                        var role = AssayService.get().getPropertyInputLineageRole(context.first);
-                        lineageRoles.add(role);
-                        materialInputs.putIfAbsent(expMaterial, role);
-                    }
-                }
-                catch (ValidationException v)
-                {
-                    // TODO: Is there a way to cancel out of this forEachResult so we stop processing?
-                    errors.addRowError(v);
-                }
-            }
-        });
-
-        if (errors.hasErrors())
-            return new InputContext(emptyMap(), emptySet());
-
-        return new InputContext(materialInputs, lineageRoles);
-    }
 }
