+++ api/src/org/labkey/api/assay/AbstractAssayTsvDataHandler.java
@@ -23,6 +23,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.json.JSONArray;
 import org.labkey.api.assay.plate.AssayPlateMetadataService;
+import org.labkey.api.assay.sample.AssaySampleLookupContext;
 import org.labkey.api.collections.CaseInsensitiveHashMap;
 import org.labkey.api.collections.CaseInsensitiveHashSet;
 import org.labkey.api.collections.Sets;
@@ -730,15 +731,15 @@ public abstract class AbstractAssayTsvDataHandler extends AbstractExperimentData
             }
             else
             {
-                ExpSampleType st = ExperimentService.get().getLookupSampleType(pd, container, user);
-                if (st != null)
+                var sampleLookup = AssaySampleLookupContext.checkSampleLookup(container, user, pd);
+                if (sampleLookup.expSampleType() != null)
                 {
                     if (pd.getPropertyType().getJdbcType().isText())
-                        lookupToSampleTypeByName.put(pd, st);
+                        lookupToSampleTypeByName.put(pd, sampleLookup.expSampleType());
                     else
-                        lookupToSampleTypeById.put(pd, st);
+                        lookupToSampleTypeById.put(pd, sampleLookup.expSampleType());
                 }
-                else if (ExperimentService.get().isLookupToMaterials(pd))
+                else if (sampleLookup.isLookup())
                 {
                     if (pd.getPropertyType().getJdbcType().isText())
                         lookupToAllSamplesByName.add(pd);
